Refactored buffers, shaders, and renderers

+++ OpenGLTest2/src/math/utils.cpp
@@ -3,7 +3,7 @@
 namespace VR { namespace math {
 
 float Utils::toRadians(float radians) {
-	return radians * PI / 180.0f;
+	return (float)radians * PI / 180.0f;
 }
 
 } }
